% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strMNCI.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{R program code for the computation of the Miettinen-Nurminen confidence limits for stratified 2x2 data
by Bernhard Klingenberg, Department of Mathematics & Statistics, Williams College,Williamtown, MA, USA, 31.8.2012.
For the code to work, each row has to contain y1, n1, y2 and n2, in that order,
where y1 is the binomial response out of n1 trials for the treatment group and y2 is
the binomial response out of n2 trials for the control group. Each row presents a
stratum. The procedure then computes the CI for the common difference in the
proportions of the treated group minus the control group.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 4 columns.
}
\usage{
dataset
}
\description{
Below are the necessary functions, the very last one is the main function that you
call to get the CI. E.g., if we had three strata
}
\keyword{datasets}
