% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MN.weights.R
\name{MN.weights}
\alias{MN.weights}
\title{computes the MN wights necessary for CI for stratified data}
\usage{
MN.weights(delta, dataset, bias.correct = FALSE)
}
\arguments{
\item{delta}{}

\item{dataset}{dataset needs to have centers as rows and number of successes in group1, number of trials in group 1, number of successes in group2, number of trials in group 2 as columns.}

\item{bias.correct}{Bessel's correction for a finite sample.}
}
\description{
computes the MN wights necessary for CI for stratified data
}
\references{
Cochran-Mantel-Haenszel Weighted Miettinen & Nurminen Method for Confidence Intervals of the Difference in Binomial Proportions from Stratified 2x2 Samples, Kaifeng Lu (2008)
}
