% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resMLE.MN.R
\name{resMLE.MN}
\alias{resMLE.MN}
\title{gives MLEs for p1,p2 under restriction p1-p2=delta}
\usage{
resMLE.MN(delta, y, n)
}
\arguments{
\item{delta}{risk difference(Restriction)}

\item{y}{the number of cases (Restriction)}

\item{n}{the number of subjects (Restriction)}
}
\description{
gives MLEs for p1,p2 under restriction p1-p2=delta
}
\references{
Eqs.27 and 28 in "Comparative analysis of two rates", MIETTINEN AND NURMINEN, Stat. Med.(1985)
}
