% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resMLE.MN1.R
\name{resMLE.MN1}
\alias{resMLE.MN1}
\title{computes restricted MLEs for pi1, pi2 under restriction pi1-pi2=delta
for a given 2x2 table and returns components
of MN chi-square function}
\usage{
resMLE.MN1(delta, y, n)
}
\description{
computes restricted MLEs for pi1, pi2 under restriction pi1-pi2=delta
for a given 2x2 table and returns components
of MN chi-square function
}
