% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreint.strat.MN.R
\name{scoreint.strat.MN}
\alias{scoreint.strat.MN}
\title{computes the CI for a common risk difference from several stratified 2x2 tables}
\usage{
scoreint.strat.MN(dataset, weight = "MN", conflev = 0.95, bias.correct = FALSE)
}
\arguments{
\item{dataset}{a matrix-class object with each row has to contain y1, n1, y2 and n2, in that order, where y1 is the binomial response out of n1 trials for the treatment group and y2 is the binomial response out of n2 trials for the control group. Each row presents a stratum.}

\item{weight}{"inv.Var", "Add4" or "Cochran". MN: "Miettinen Nurminen (1985)" The solution for "scorestat.MN" via uniroot().}

\item{conflev}{}

\item{bias.correct}{Bessel's correction for a finite sample.}
}
\description{
computes the CI for a common risk difference from several stratified 2x2 tables
}
\references{
Miettinen And Nurminen 1985, Stat. in Medicine
}
