% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorestat.MN.R
\name{scorestat.MN}
\alias{scorestat.MN}
\title{computes the chi-square function}
\usage{
scorestat.MN(delta0, dataset, weight = "MN", conflev, bias.correct = FALSE)
}
\arguments{
\item{delta0}{a given null value}

\item{dataset}{stratified 2x2 tables, needs to have centers as rows. number of successes in group1, number of trials in group 1,  number of successes in group2, number of trials in group 2}

\item{weight}{Weight specification.}

\item{conflev}{Confidence level}

\item{bias.correct}{Bessel's correction for a finite sample.}
}
\description{
computes the chi-square function
}
\references{
Miettinen and Nurminen 1985, Stat. in Medicine, page 218
}
